Fixed WritePackets Function and updated Read function

+++ app/src/main/java/me/iologic/apps/dtn/FileServices.java
@@ -92,7 +92,7 @@ public class FileServices {
         return ret;
     }
 
-    public float getFileSize(){
+    public double getFileSize(){
         return fileContent.length();
     }
